# Remove JSON building in favor to use built-in types