--- a/.github/actions/legacy_unit_tests.sh
+++ b/.github/actions/legacy_unit_tests.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 clean() {
-    echo "Cleaning build agent and dependencies..."
+    echo "Cleaning build and dependencies..."
     cd "$GITHUB_WORKSPACE"
     make clean-deps -C src
     rm -rf src/external/*
@@ -19,10 +19,12 @@ build_wazuh_test_flags() {
 
 build_wazuh_unit_tests() {
     local target=$1
-    echo "Building Wazuh Agent Unit Tests for target: $target"
+    echo "Building Wazuh Unit Tests for target: $target"
     mkdir -p "$GITHUB_WORKSPACE/src/unit_tests/build"
     cd "$GITHUB_WORKSPACE/src/unit_tests/build"
     if [[ $target == "agent" ]]; then
+        cmake -DTARGET=${target} ..
+    elif [[ $target == "server" ]]; then
         cmake -DTARGET=${target} ..
     elif [[ $target == "winagent" ]]; then
         cmake -DTARGET=${target} -DCMAKE_TOOLCHAIN_FILE=../Toolchain-win32.cmake ..
@@ -32,9 +34,12 @@ build_wazuh_unit_tests() {
 
 run_wazuh_unit_tests() {
     local target=$1
-    echo "Running Wazuh Agent Unit Tests for target: $target"
+    echo "Running Wazuh Unit Tests for target: $target"
     cd "$GITHUB_WORKSPACE/src/unit_tests/build"
     if [[ $target == "agent" ]]; then
+        ctest --output-on-failure  > "test_results.txt" || true
+        make coverage > "coverage_results.txt" || true
+    elif [[ $target == "server" ]]; then
         ctest --output-on-failure  > "test_results.txt" || true
         make coverage > "coverage_results.txt" || true
     elif [[ $target == "winagent" ]]; then
@@ -55,7 +60,7 @@ format_display_test_results() {
 
 
 format_display_test_coverage() {
-    if [[ $1 == "agent" ]]; then
+    if [[ $1 != "winagent" ]]; then
         echo "Formatting and displaying test coverage"
         cd "$GITHUB_WORKSPACE/src/unit_tests/build"
         if grep "Summary coverage rate:" coverage_results.txt > /dev/null; then
@@ -74,8 +79,8 @@ main() {
         exit 1
     fi
 
-    if [[ $target != "agent" && $target != "winagent" ]]; then
-        echo "Error: Invalid target '$target'. Expected 'agent' or 'winagent'."
+    if [[ $target != "agent" && $target != "winagent" && $target != "server" ]]; then
+        echo "Error: Invalid target '$target'. Expected 'agent', 'winagent' or 'server'."
         exit 1
     fi