--- a/src/unit_tests/wazuh_db/test_wdb_global.c
+++ b/src/unit_tests/wazuh_db/test_wdb_global.c
@@ -170,17 +170,13 @@ void create_wdb_global_unassign_agent_group_success_call(int agent_id, int group
 /**
  * @brief Configure all the wrappers to simulate a successful call to wdb_global_calculate_agent_group_csv() method
  * @param agent_id The id of the agent to calculate its groups csv
- * @param find_group_resp The response of the find select group query
  */
-void create_wdb_global_calculate_agent_group_csv_success_call(int agent_id, cJSON* j_group_resp) {
+void create_wdb_global_calculate_agent_group_csv_success_call(int agent_id) {
     will_return(__wrap_wdb_begin2, 1);
     will_return(__wrap_wdb_stmt_cache, OS_SUCCESS);
     expect_value(__wrap_sqlite3_bind_int, index, 1);
     expect_value(__wrap_sqlite3_bind_int, value, agent_id);
     will_return(__wrap_sqlite3_bind_int, SQLITE_OK);
-    /* wdb_exec_stmt_sized */
-    wrap_wdb_exec_stmt_sized_success_call(j_group_resp, STMT_SINGLE_COLUMN);
-    expect_function_call(__wrap_cJSON_Delete);
 }
 
 /**
@@ -5826,8 +5822,6 @@ void test_wdb_global_delete_group_success(void **state)
     cJSON *sql_agents_id = cJSON_Parse("[{\"id_agent\":1}]");
     int agent_id = 1;
     cJSON* j_priority_resp = cJSON_Parse("[{\"id\":0}]");
-    cJSON* j_group_array = __real_cJSON_CreateArray();
-    cJSON_AddItemToArray(j_group_array, cJSON_CreateString(group_name));
     char hash[] = "19dcd0dd"; //"GROUP" hash
 
     //wdb_is_group_empty
@@ -5840,7 +5834,6 @@ void test_wdb_global_delete_group_success(void **state)
 
     will_return(__wrap_wdb_begin2, 1);
     will_return(__wrap_wdb_stmt_cache, 1);
-
     expect_value(__wrap_sqlite3_bind_text, pos, 1);
     expect_string(__wrap_sqlite3_bind_text, buffer, group_name);
     will_return(__wrap_sqlite3_bind_text, SQLITE_OK);
@@ -5853,7 +5846,12 @@ void test_wdb_global_delete_group_success(void **state)
     create_wdb_global_get_agent_max_group_priority_success_call(agent_id, j_priority_resp);
 
     /* wdb_global_calculate_agent_group_csv */
-    create_wdb_global_calculate_agent_group_csv_success_call(agent_id, j_group_array);
+    create_wdb_global_calculate_agent_group_csv_success_call(agent_id);
+
+    will_return(__wrap_wdb_step, SQLITE_ROW);
+    expect_value(__wrap_sqlite3_column_text, iCol, 0);
+    will_return(__wrap_sqlite3_column_text, group_name);
+    will_return(__wrap_wdb_step, SQLITE_DONE);
 
     /* wdb_global_set_agent_group_context */
     create_wdb_global_set_agent_group_context_success_call(agent_id, group_name, hash, sync_status);
@@ -5864,7 +5862,6 @@ void test_wdb_global_delete_group_success(void **state)
 
     assert_int_equal(result, OS_SUCCESS);
     __real_cJSON_Delete(j_priority_resp);
-    __real_cJSON_Delete(j_group_array);
     __real_cJSON_Delete(sql_agents_id);
 }
 
@@ -7760,10 +7757,8 @@ void test_wdb_global_calculate_agent_group_csv_unable_to_get_group(void **state)
 
     // wdb_global_select_group_belong
     data->wdb->transaction = 0;
-    will_return(__wrap_wdb_begin2, OS_INVALID);
     expect_string(__wrap__mdebug1, formatted_msg, "Cannot begin transaction");
-
-    expect_string(__wrap__mdebug1, formatted_msg, "Unable to get groups of agent '001'");
+    will_return(__wrap_wdb_begin2, OS_INVALID);
 
     char *result = wdb_global_calculate_agent_group_csv(data->wdb, agent_id);
 
@@ -7774,7 +7769,6 @@ void test_wdb_global_calculate_agent_group_csv_success(void **state) {
     test_struct_t *data  = (test_struct_t *)*state;
     int agent_id = 1;
 
-    // wdb_global_select_group_belong
     data->wdb->transaction = 1;
     will_return(__wrap_wdb_stmt_cache, OS_SUCCESS);
     expect_value(__wrap_sqlite3_bind_int, index, 1);
@@ -7783,10 +7777,14 @@ void test_wdb_global_calculate_agent_group_csv_success(void **state) {
     cJSON *j_groups = __real_cJSON_CreateArray();
     cJSON_AddItemToArray(j_groups, cJSON_CreateString("group1"));
     cJSON_AddItemToArray(j_groups, cJSON_CreateString("group2"));
-    /* wdb_exec_stmt_sized */
-    wrap_wdb_exec_stmt_sized_success_call(j_groups, STMT_SINGLE_COLUMN);
 
-    expect_function_call(__wrap_cJSON_Delete);
+    will_return(__wrap_wdb_step, SQLITE_ROW);
+    expect_value(__wrap_sqlite3_column_text, iCol, 0);
+    will_return(__wrap_sqlite3_column_text, "group1");
+    will_return(__wrap_wdb_step, SQLITE_ROW);
+    expect_value(__wrap_sqlite3_column_text, iCol, 0);
+    will_return(__wrap_sqlite3_column_text, "group2");
+    will_return(__wrap_wdb_step, SQLITE_DONE);
 
     char *result = wdb_global_calculate_agent_group_csv(data->wdb, agent_id);
 
@@ -8628,7 +8626,12 @@ void test_wdb_global_set_agent_groups_override_success(void **state) {
         create_wdb_global_assign_agent_group_success_call(agent_id, group_id, group_name, j_find_group_resp);
 
         /* wdb_global_calculate_agent_group_csv */
-        create_wdb_global_calculate_agent_group_csv_success_call(agent_id, j_group_array);
+        create_wdb_global_calculate_agent_group_csv_success_call(agent_id);
+
+        will_return(__wrap_wdb_step, SQLITE_ROW);
+        expect_value(__wrap_sqlite3_column_text, iCol, 0);
+        will_return(__wrap_sqlite3_column_text, group_name);
+        will_return(__wrap_wdb_step, SQLITE_DONE);
 
         /* wdb_global_set_agent_group_context */
         create_wdb_global_set_agent_group_context_success_call(agent_id, group_name, hash, sync_status);
@@ -8666,6 +8669,7 @@ void test_wdb_global_set_agent_groups_override_delete_error(void **state) {
         /* wdb_global_delete_agent_belong */
         will_return(__wrap_wdb_begin2, -1);
         expect_string(__wrap__mdebug1, formatted_msg, "Cannot begin transaction");
+
         expect_string(__wrap__merror, formatted_msg, "There was an error cleaning the previous agent groups");
 
         /* wdb_global_validate_groups_success_call */
@@ -8675,7 +8679,12 @@ void test_wdb_global_set_agent_groups_override_delete_error(void **state) {
         create_wdb_global_assign_agent_group_success_call(agent_id, group_id, group_name, j_find_group_resp);
 
         /* wdb_global_calculate_agent_group_csv */
-        create_wdb_global_calculate_agent_group_csv_success_call(agent_id, j_group_array);
+        create_wdb_global_calculate_agent_group_csv_success_call(agent_id);
+
+        will_return(__wrap_wdb_step, SQLITE_ROW);
+        expect_value(__wrap_sqlite3_column_text, iCol, 0);
+        will_return(__wrap_sqlite3_column_text, group_name);
+        will_return(__wrap_wdb_step, SQLITE_DONE);
 
         /* wdb_global_set_agent_group_context */
         create_wdb_global_set_agent_group_context_success_call(agent_id, group_name, hash, sync_status);
@@ -8722,7 +8731,12 @@ void test_wdb_global_set_agent_groups_add_modes_assign_error(void **state) {
         expect_string(__wrap__merror, formatted_msg, "There was an error assigning the groups to agent '001'");
 
         /* wdb_global_calculate_agent_group_csv */
-        create_wdb_global_calculate_agent_group_csv_success_call(agent_id, j_group_array);
+        create_wdb_global_calculate_agent_group_csv_success_call(agent_id);
+
+        will_return(__wrap_wdb_step, SQLITE_ROW);
+        expect_value(__wrap_sqlite3_column_text, iCol, 0);
+        will_return(__wrap_sqlite3_column_text, group_name);
+        will_return(__wrap_wdb_step, SQLITE_DONE);
 
         /* wdb_global_set_agent_group_context */
         create_wdb_global_set_agent_group_context_success_call(agent_id, group_name, hash, sync_status);
@@ -8769,7 +8783,12 @@ void test_wdb_global_set_agent_groups_append_success(void **state) {
         create_wdb_global_assign_agent_group_success_call(agent_id, group_id, group_name, j_find_group_resp);
 
         /* wdb_global_calculate_agent_group_csv */
-        create_wdb_global_calculate_agent_group_csv_success_call(agent_id, j_group_array);
+        create_wdb_global_calculate_agent_group_csv_success_call(agent_id);
+
+        will_return(__wrap_wdb_step, SQLITE_ROW);
+        expect_value(__wrap_sqlite3_column_text, iCol, 0);
+        will_return(__wrap_sqlite3_column_text, group_name);
+        will_return(__wrap_wdb_step, SQLITE_DONE);
 
         /* wdb_global_set_agent_group_context */
         create_wdb_global_set_agent_group_context_success_call(agent_id, group_name, hash, sync_status);
@@ -8816,7 +8835,12 @@ void test_wdb_global_set_agent_groups_empty_only_success(void **state) {
         create_wdb_global_assign_agent_group_success_call(agent_id, group_id, group_name, j_find_group_resp);
 
         /* wdb_global_calculate_agent_group_csv */
-        create_wdb_global_calculate_agent_group_csv_success_call(agent_id, j_group_array);
+        create_wdb_global_calculate_agent_group_csv_success_call(agent_id);
+
+        will_return(__wrap_wdb_step, SQLITE_ROW);
+        expect_value(__wrap_sqlite3_column_text, iCol, 0);
+        will_return(__wrap_sqlite3_column_text, group_name);
+        will_return(__wrap_wdb_step, SQLITE_DONE);
 
         /* wdb_global_set_agent_group_context */
         create_wdb_global_set_agent_group_context_success_call(agent_id, group_name, hash, sync_status);
@@ -8889,7 +8913,12 @@ void test_wdb_global_set_agent_groups_remove_success(void **state) {
         create_wdb_global_get_agent_max_group_priority_success_call(agent_id, j_priority_resp);
 
         /* wdb_global_calculate_agent_group_csv */
-        create_wdb_global_calculate_agent_group_csv_success_call(agent_id, j_group_array);
+        create_wdb_global_calculate_agent_group_csv_success_call(agent_id);
+
+        will_return(__wrap_wdb_step, SQLITE_ROW);
+        expect_value(__wrap_sqlite3_column_text, iCol, 0);
+        will_return(__wrap_sqlite3_column_text, group_name);
+        will_return(__wrap_wdb_step, SQLITE_DONE);
 
         /* wdb_global_set_agent_group_context */
         create_wdb_global_set_agent_group_context_success_call(agent_id, group_name, hash, sync_status);
@@ -8929,7 +8958,12 @@ void test_wdb_global_set_agent_groups_remove_unassign_error(void **state) {
         expect_string(__wrap__merror, formatted_msg, "There was an error un-assigning the groups to agent '001'");
 
         /* wdb_global_calculate_agent_group_csv */
-        create_wdb_global_calculate_agent_group_csv_success_call(agent_id, j_group_array);
+        create_wdb_global_calculate_agent_group_csv_success_call(agent_id);
+
+        will_return(__wrap_wdb_step, SQLITE_ROW);
+        expect_value(__wrap_sqlite3_column_text, iCol, 0);
+        will_return(__wrap_sqlite3_column_text, group_name);
+        will_return(__wrap_wdb_step, SQLITE_DONE);
 
         /* wdb_global_set_agent_group_context */
         create_wdb_global_set_agent_group_context_success_call(agent_id, group_name, hash, sync_status);
@@ -8993,7 +9027,6 @@ void test_wdb_global_set_agent_groups_calculate_csv_empty(void **state) {
         /* wdb_global_calculate_agent_group_csv */
         will_return(__wrap_wdb_begin2, -1);
         expect_string(__wrap__mdebug1, formatted_msg, "Cannot begin transaction");
-        expect_string(__wrap__mdebug1, formatted_msg, "Unable to get groups of agent '001'");
         expect_string(__wrap__mwarn, formatted_msg, "The groups were empty right after the set for agent '001'");
 
         /* wdb_global_set_agent_group_context */
@@ -9035,7 +9068,12 @@ void test_wdb_global_set_agent_groups_set_group_ctx_error(void **state) {
         create_wdb_global_get_agent_max_group_priority_success_call(agent_id, j_priority_resp);
 
         /* wdb_global_calculate_agent_group_csv */
-        create_wdb_global_calculate_agent_group_csv_success_call(agent_id, j_group_array);
+        create_wdb_global_calculate_agent_group_csv_success_call(agent_id);
+
+        will_return(__wrap_wdb_step, SQLITE_ROW);
+        expect_value(__wrap_sqlite3_column_text, iCol, 0);
+        will_return(__wrap_sqlite3_column_text, group_name);
+        will_return(__wrap_wdb_step, SQLITE_DONE);
 
         /* wdb_global_set_agent_group_context */
         expect_value(__wrap_wdb_init_stmt_in_cache, statement_index, WDB_STMT_GLOBAL_GROUP_CTX_SET);
@@ -9288,86 +9326,38 @@ void test_wdb_global_recalculate_all_agent_groups_hash_bind_fail(void **state)
     assert_int_equal(result, OS_INVALID);
 }
 
-void test_wdb_global_recalculate_all_agent_groups_hash_exec_stmt_null(void **state)
-{
-    test_struct_t *data  = (test_struct_t *)*state;
-
-    will_return(__wrap_wdb_begin2, 1);
-    will_return(__wrap_wdb_stmt_cache, 1);
-    expect_value(__wrap_sqlite3_bind_int, index, 1);
-    expect_value(__wrap_sqlite3_bind_int, value, 0);
-    will_return(__wrap_sqlite3_bind_int, SQLITE_OK);
-    will_return(__wrap_wdb_exec_stmt, NULL);
-
-    expect_function_call(__wrap_cJSON_Delete);
-
-    int result = wdb_global_recalculate_all_agent_groups_hash(data->wdb);
-
-    assert_int_equal(result, OS_SUCCESS);
-}
-
-void test_wdb_global_recalculate_all_agent_groups_hash_invalid_id(void **state)
-{
-    test_struct_t *data  = (test_struct_t *)*state;
-    cJSON *json_agent = NULL;
-
-    json_agent = cJSON_CreateObject();
-    cJSON_AddStringToObject(json_agent, "id", "id_str");
-
-    will_return(__wrap_wdb_begin2, 1);
-    will_return(__wrap_wdb_stmt_cache, 1);
-    expect_value(__wrap_sqlite3_bind_int, index, 1);
-    expect_value(__wrap_sqlite3_bind_int, value, 0);
-    will_return(__wrap_sqlite3_bind_int, SQLITE_OK);
-    will_return(__wrap_wdb_exec_stmt, json_agent);
-
-    expect_string(__wrap__merror, formatted_msg, "Invalid element returned by get all agents query");
-
-    expect_function_call(__wrap_cJSON_Delete);
-
-    int result = wdb_global_recalculate_all_agent_groups_hash(data->wdb);
-
-    assert_int_equal(result, OS_INVALID);
-    __real_cJSON_Delete(json_agent);
-}
-
 void test_wdb_global_recalculate_all_agent_groups_hash_recalculate_error(void **state)
 {
     test_struct_t *data  = (test_struct_t *)*state;
-    cJSON *json_agent = NULL;
     int agent_id = 1;
-
-    cJSON* j_stmt_result = __real_cJSON_CreateArray();
-    json_agent = cJSON_CreateObject();
-    cJSON_AddItemToObject(json_agent, "id", cJSON_CreateNumber(agent_id));
-    cJSON_AddItemToArray(j_stmt_result, json_agent);
+    char *group_name = "GROUP";
 
     will_return(__wrap_wdb_begin2, 1);
     will_return(__wrap_wdb_stmt_cache, 1);
     expect_value(__wrap_sqlite3_bind_int, index, 1);
     expect_value(__wrap_sqlite3_bind_int, value, 0);
     will_return(__wrap_sqlite3_bind_int, SQLITE_OK);
-    will_return(__wrap_wdb_exec_stmt, j_stmt_result);
+    will_return(__wrap_wdb_step, SQLITE_ROW);
+    expect_value(__wrap_sqlite3_column_int, iCol, 0);
+    will_return(__wrap_sqlite3_column_int, 1);
+    expect_value(__wrap_sqlite3_column_text, iCol, 1);
+    will_return(__wrap_sqlite3_column_text, group_name);
 
     /* wdb_global_calculate_agent_group_csv */
     will_return(__wrap_wdb_begin2, -1);
     expect_string(__wrap__mdebug1, formatted_msg, "Cannot begin transaction");
-    expect_string(__wrap__mdebug1, formatted_msg, "Unable to get groups of agent '001'");
     expect_string(__wrap__mdebug1, formatted_msg, "No groups in belongs table for agent '001'");
 
-    /* wdb_global_set_agent_group_context */
+    /* wdb_global_set_agent_group_hash */
     expect_value(__wrap_wdb_init_stmt_in_cache, statement_index, WDB_STMT_GLOBAL_GROUP_HASH_SET);
     will_return(__wrap_wdb_init_stmt_in_cache, NULL);
     expect_string(__wrap__merror, formatted_msg, "There was an error assigning the groups hash to agent '001'");
 
     expect_string(__wrap__merror, formatted_msg, "Couldn't recalculate hash group for agent: '001'");
 
-    expect_function_call(__wrap_cJSON_Delete);
-
     int result = wdb_global_recalculate_all_agent_groups_hash(data->wdb);
 
     assert_int_equal(result, OS_INVALID);
-    __real_cJSON_Delete(j_stmt_result);
 }
 
 
@@ -9847,8 +9837,6 @@ int main()
         cmocka_unit_test_setup_teardown(test_wdb_global_recalculate_all_agent_groups_hash_transaction_fail, test_setup, test_teardown),
         cmocka_unit_test_setup_teardown(test_wdb_global_recalculate_all_agent_groups_hash_cache_fail, test_setup, test_teardown),
         cmocka_unit_test_setup_teardown(test_wdb_global_recalculate_all_agent_groups_hash_bind_fail, test_setup, test_teardown),
-        cmocka_unit_test_setup_teardown(test_wdb_global_recalculate_all_agent_groups_hash_exec_stmt_null, test_setup, test_teardown),
-        cmocka_unit_test_setup_teardown(test_wdb_global_recalculate_all_agent_groups_hash_invalid_id, test_setup, test_teardown),
         cmocka_unit_test_setup_teardown(test_wdb_global_recalculate_all_agent_groups_hash_recalculate_error, test_setup, test_teardown),
     };