… .github/workflows/legacy-unit-tests.yml
… src/unit_tests/wazuh_db/CMakeLists.txt