|Related issue|
|---|
|Closes https://github.com/wazuh/devel-major-incidents/issues/15|

## Description

This PR aims to use the built-in types in favour the get better performance results in the usage.
In other side, if we get the current hash, we can check if after the calculation the number change, to decide if the update query will be processed.

### Results previous the change
2079149 / 246 -> 8451 ms
### Results after the change
427208 / 2197 -> 194 ms