@MiguelazoDS 2025-03-25
@Dwordcito if we do not want too much changes in this implementation since it was already tested in the failing environment, we can just dismiss the discussion above if you agree @pereyra-m